# openapi.yaml
openapi: 3.0.0
info:
  title: Leanstock API
  description: Production-grade inventory management system with multi-tenant architecture
  version: 1.0.0
  contact:
    name: Leanstock Team
    email: support@leanstock.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.leanstock.com/api/v1
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: List all products with pagination
      description: Get paginated list of products for the tenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Products
      summary: Create a new product
      description: Create a new product (requires products:create permission)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            examples:
              new_product:
                summary: New product creation
                value:
                  sku: "PROD-001"
                  name: "Premium Widget"
                  description: "High-quality widget for industrial use"
                  baseCost: 25.50
                  retailPrice: 49.99
                  weight: 1.5
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /inventory:
    get:
      tags:
        - Inventory
      summary: List all inventory with filters
      description: Get paginated list of inventory records with filtering options
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - name: productId
          in: query
          description: Filter by product ID
          required: false
          schema:
            type: string
        - name: locationId
          in: query
          description: Filter by location ID
          required: false
          schema:
            type: string
        - name: lowStock
          in: query
          description: Filter low stock items
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Inventory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Inventory
      summary: Create inventory record
      description: Create new inventory record (requires inventory:create permission)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryRequest'
      responses:
        '201':
          description: Inventory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - baseCost
        - retailPrice
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 50
          example: "PROD-001"
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: "Premium Widget"
        description:
          type: string
          maxLength: 1000
          example: "High-quality widget for industrial use"
        baseCost:
          type: number
          minimum: 0
          example: 25.50
        retailPrice:
          type: number
          minimum: 0
          example: 49.99
        weight:
          type: number
          minimum: 0
          example: 1.5
...
    CreateInventoryRequest:
      type: object
      required:
        - productId
        - locationId
        - quantity
      properties:
        productId:
          type: string
          example: "prod-123"
        locationId:
          type: string
          example: "loc-123"
        quantity:
          type: integer
          minimum: 0
          example: 100
        reservedQuantity:
          type: integer
          minimum: 0
          default: 0
          example: 10
        inTransit:
          type: integer
          minimum: 0
          default: 0
          example: 5
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: "prod-123"
        sku:
          type: string
          example: "PROD-001"
        name:
          type: string
          example: "Premium Widget"
        description:
          type: string
          example: "High-quality widget for industrial use"
        baseCost:
          type: number
          example: 25.50
        retailPrice:
          type: number
          example: 49.99
        weight:
          type: number
          example: 1.5
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
...
    PaginatedProductsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    InventoryResponse:
      type: object
      properties:
        id:
          type: string
          example: "inv-123"
        productId:
          type: string
          example: "prod-123"
        locationId:
          type: string
          example: "loc-123"
        quantity:
          type: integer
          example: 100
        reservedQuantity:
          type: integer
          example: 10
        inTransit:
          type: integer
          example: 5
        daysInInventory:
          type: integer
          example: 45
        lastMovedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        product:
          $ref: '#/components/schemas/ProductResponse'
        location:
          $ref: '#/components/schemas/LocationResponse'
    PaginatedInventoryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/InventoryResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    LocationResponse:
      type: object
      properties:
        id:
          type: string
          example: "loc-123"
        name:
          type: string
          example: "Main Warehouse"
        type:
          type: string
          example: "WAREHOUSE"
        address:
          type: string
          example: "123 Main St, City, State"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # Error schemas
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    SearchParam:
      name: search
      in: query
      description: Search term for filtering
      required: false
      schema:
        type: string
        maxLength: 100
      example: "widget"
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      required: false
      schema:
        type: string
        enum: [name, sku, createdAt, updatedAt]
        default: createdAt
      example: name
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      example: asc
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            bad_request:
              summary: Bad request example
              value:
                error: "Bad Request"
                message: "Invalid request data"
                details: []
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Unauthorized example
              value:
                error: "Unauthorized"
                message: "Authentication required"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: Forbidden example
              value:
                error: "Forbidden"
                message: "Insufficient permissions"
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              summary: Conflict example
              value:
                error: "Conflict"
                message: "Resource already exists"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              summary: Validation error example
              value:
                error: "Validation Error"
                message: "Input validation failed"
                details:
                  - field: "email"
                    message: "Invalid email format"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              summary: Internal server error example
              value:
                error: "Internal Server Error"
                message: "An unexpected error occurred"
...